Moved versionHandler and related structs in utils package
